requirements.txt
requests>=2.2.3
python-dotenv>=1.2.0
terminaltables>=3.1.10